requirements.txt
numpy~=1.18.5
pandas~=1.1.0
tensorflow-gpu~=2.3.0